TSFF-1445: Lager egen dto for opphørsdato, maksdato og antall dager forbrukt (#336)
* TSFF-1445: Lager egen dto for opphørsdato, maksdato og antall dager forbrukt

* TSFF-1445: Fikser path

* ung-sak.openapi.json updated by build pipeline 


skip-checks:true

---------

+++ domenetjenester/ungdomsprogram/src/main/java/no/nav/ung/sak/ungdomsprogram/forbruktedager/FinnForbrukteDager.java
@@ -18,7 +18,7 @@ public class FinnForbrukteDager {
 
         var kunVirkedager = ungdomsprogramperiode.disjoint(helger);
 
-        long antallDager = 0;
+        int antallDager = 0;
         LocalDateTimeline<Boolean> resultatTidslinje = LocalDateTimeline.empty();
         for (LocalDateSegment<Boolean> virkedagSegment : kunVirkedager.toSegments()) {
             var antallDagerISegment = virkedagSegment.getLocalDateInterval().totalDays();
@@ -27,7 +27,7 @@ public class FinnForbrukteDager {
             }
             if (antallDagerISegment + antallDager < MAKS_ANTALL_DAGER) {
                 resultatTidslinje = resultatTidslinje.crossJoin(new LocalDateTimeline<>(List.of(virkedagSegment)));
-                antallDager += antallDagerISegment;
+                antallDager += (int) antallDagerISegment;
             } else {
                 var delAvPeriode = finnDelAvPeriodeSomInnvilges(virkedagSegment, antallDager);
                 resultatTidslinje = resultatTidslinje.crossJoin(delAvPeriode);
@@ -48,7 +48,7 @@ public class FinnForbrukteDager {
         return new VurderAntallDagerResultat(LocalDateTimeline.empty(), antallDager);
     }
 
-    private static LocalDateTimeline<Boolean> finnDelAvPeriodeSomInnvilges(LocalDateSegment<Boolean> virkedagSegment, long antallDager) {
+    private static LocalDateTimeline<Boolean> finnDelAvPeriodeSomInnvilges(LocalDateSegment<Boolean> virkedagSegment, int antallDager) {
         var dagerSomGjenstår = MAKS_ANTALL_DAGER - antallDager;
 
         if (dagerSomGjenstår < 1) {
