+++ domenetjenester/ungdomsprogram/src/main/java/no/nav/ung/sak/ungdomsprogram/forbruktedager/VurderAntallDagerResultat.java
@@ -2,5 +2,5 @@ package no.nav.ung.sak.ungdomsprogram.forbruktedager;
 
 import no.nav.fpsak.tidsserie.LocalDateTimeline;
 
-public record VurderAntallDagerResultat(LocalDateTimeline<Boolean> tidslinjeNokDager, long forbrukteDager) {
+public record VurderAntallDagerResultat(LocalDateTimeline<Boolean> tidslinjeNokDager, int forbrukteDager) {
 }
